Fix Emscripten/WASM build: resolve all compile and link errors
- cmake/linux/compilerconfig.cmake: guard -march=x86-64-v2/-native behind
  !Emscripten check
- thirdparty/CMakeLists.txt: make sljit INTERFACE library on Emscripten
- nall/intrinsics.hpp: add WASM32 platform/architecture/endian detection
- nall/float-env.hpp: add WASM stub (all FP env constants = 0)
- nall/recompiler/generic/generic.hpp: add stub struct for non-SLJIT builds
- nall/sljitAllocator.cpp: guard with !__EMSCRIPTEN__
- nall/thread.hpp: guard pthread_setname_np with !__EMSCRIPTEN__
- ares/ares/ares.hpp: guard #include <sljit.h> with !__EMSCRIPTEN__
- ares/n64/cpu/interpreter-fpu.cpp: guard duplicate switch cases on WASM
- ares/n64/cpu/recompiler.cpp: wrap entire file with !__EMSCRIPTEN__
- ares/n64/rsp/recompiler.cpp: wrap entire file with !__EMSCRIPTEN__
- libco/emscripten.c: fix co_create signature (3-arg → 2-arg)
- mia/medium/famicom.cpp, mia/pak/pak.cpp: cast u64→size_t to fix
  narrowing errors on wasm32
- web/CMakeLists.txt: use --use-port=emdawnwebgpu, add nall + nall
  include path, MIA_LIBRARY define, rdp-cpu-parser.cpp source
- web/webgpu-rdp.cpp: update to Dawn emdawnwebgpu API
  (WGPUShaderSourceWGSL, WGPUStringView, WGPUBufferUsage,
   wgpuBufferMapAsync new signature, ::RDP:: qualified names)
- web/platform-web.hpp: remove invalid 'override' on non-virtual dtor
- web/platform-web.cpp: fix node->parent() weak_ptr usage (.lock())
- web/main-web.cpp: move s_frameCount declaration before first use

Build produces build_wasm/web/ares-web.js + ares-web.wasm successfully.

+++ mia/medium/famicom.cpp
@@ -40,23 +40,23 @@ auto Famicom::load(string location) -> LoadResult {
 
   std::span<const u8> view{rom};
   if(auto node = document["game/board/memory(type=ROM,content=iNES)"]) {
-    pak->append("ines.rom", {view.data(), node["size"].natural()});
+    pak->append("ines.rom", {view.data(), (size_t)node["size"].natural()});
     view = view.subspan(node["size"].natural());
   }
   if(auto node = document["game/board/memory(type=Flash,content=Program)"]) {
-    pak->append("program.flash", {view.data(), node["size"].natural()});
+    pak->append("program.flash", {view.data(), (size_t)node["size"].natural()});
     Pak::load("program.flash", ".flash");
     view = view.subspan(node["size"].natural());
   } else if(auto node = document["game/board/memory(type=ROM,content=Program)"]) {
-    pak->append("program.rom", {view.data(), node["size"].natural()});
+    pak->append("program.rom", {view.data(), (size_t)node["size"].natural()});
     view = view.subspan(node["size"].natural());
   }
   if(auto node = document["game/board/memory(type=ROM,content=Option)"]) {
-    pak->append("option.rom", {view.data(), node["size"].natural()});
+    pak->append("option.rom", {view.data(), (size_t)node["size"].natural()});
     view = view.subspan(node["size"].natural());
   }
   if(auto node = document["game/board/memory(type=ROM,content=Character)"]) {
-    pak->append("character.rom", {view.data(), node["size"].natural()});
+    pak->append("character.rom", {view.data(), (size_t)node["size"].natural()});
     view = view.subspan(node["size"].natural());
   }
